requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
fastapi==0.74.1
h11==0.13.0
httptools==0.3.0
idna==3.3
packaging==21.3
pydantic==1.9.0
pyparsing==3.0.7
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
redis==4.1.4
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.5
watchgod==0.7
websockets==10.2
wrapt==1.13.3
